// open_claw/agent_team_plan/ccn2_workspace/src/client/elemental-hunter/types.ts
export enum ArtifactType {
  Swap = 'Swap',
  Change = 'Change',
  Charge = 'Charge'
}

export enum ComboType {
  C3 = 'C3',
  C4 = 'C4'
}

export enum PlayerId {
  P1 = 'P1',
  P2 = 'P2'
}